requirements.txt
python-telegram-bot==21.10
google-genai==0.6.0
python-dotenv